move some base attributes and methods in simple http client to base http client, which make SimpleHttpClient more focus on network sutff. add support to add query string in simple http client
